requirements.txt
black==19.10b0
Django==3.0.6
djangorestframework==3.11.0